point to the new jsoncdc
this commit removes all references to the (now older) jsoncdc. pointing
to the new source repo "instructure/jsoncdc". it also installs the
latest version of jsoncdc which is available now on the "testing" channel
instead of unstable.

this shouldn't break anything with lapidus seeing as how the two changes are:
  - We release DEBs now (no impact).
  - We have a slightly longer rust line length (If this had an impact I'd
      be very confused).
  - We've added extra info to `commits`, but haven't removed any existing
    properties.
      (Perhaps in the future lapidus should pick up these properties?).